Extraindo interface para Mutex
